23 arpil lot of struggle

+++ backend/src/db/queries/project.query.ts
@@ -1,10 +1,10 @@
 import { pool } from '../pool.js';
 
-export const createProjectQuery = async (name: string, workspace_id: number) => {
+export const createProjectQuery = async (name: string, workspace_id: number, owner_id: number) => {
   try {
     const result = await pool.query(
-      "INSERT INTO projects (name,workspace_id) values ($1,$2) RETURNING *",
-      [name, workspace_id]
+      "INSERT INTO projects (name, workspace_id) SELECT $1, $2 WHERE EXISTS (SELECT 1 FROM workspace WHERE id = $2 AND owner_id = $3) RETURNING *",
+      [name, workspace_id, owner_id]
     );
     return result.rows[0];
   } catch (error) {
@@ -13,7 +13,7 @@ export const createProjectQuery = async (name: string, workspace_id: number) => 
   }
 };
 
-export const getAllProjectsQuery = async (workspace_id: string) => {
+export const getAllProjectsQuery = async (workspace_id: number) => {
   try {
     const result = await pool.query(
       "SELECT * FROM projects WHERE workspace_id = $1",
@@ -39,15 +39,15 @@ export const getProjectByIdQuery = async (id: string) => {
   }
 };
 
-export const deleteProjectQuery = async (id: string) => {
+export const deleteProjectQuery = async (id: string, owner_id: number) => {
   try {
     const result = await pool.query(
-      "DELETE FROM projects WHERE id=$1 RETURNING *",
-      [id]
+      "DELETE FROM projects WHERE id=$1 AND workspace_id IN (SELECT id FROM workspace WHERE owner_id=$2) RETURNING *",
+      [id, owner_id]
     );
     return result.rows[0];
   } catch (error) {
     console.error("Error in deleteProjectQuery:", error);
     throw error;
   }
-};
+};
